requirements.txt
Django>=5.0
djangorestframework>=3.14
channels>=4.0
channels-redis>=4.0
uvicorn[standard]>=0.23
Pillow>=10.0
django-cors-headers>=4.0
redis>=5.0
